dynamically functions for each character head
Prior to this commit, we had no way of producing a list of possible
words for a sorted character head; i.e. a sorted list of the characters
in the word. This list will include anagrams like `%{'em' => ["em",
"me"]}`. At compile-time we generate a static function for each
character head that statically points to the list of possible words,
thus ensuring that the dictionary is only consulted at compile-time, and
never at run time. An `@external_resource` has been established so that
if the dictionary changes, mix will know to recompile the `Robert`
module.

Here's an example of what it looks like in action:
```
Interactive Elixir (1.0.4) - press Ctrl+C to exit (type h() ENTER for
help)
iex(1)> Robert.possible_words('em')
["me", "em"]
iex(2)>
```